pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oauth2-llavemx"
version = "3.2.1"
description = "LlaveMX OAuth2 backend for Open edX using python-social-auth."
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [{ name = "EMI / MéxicoX", email = "soporte@mexicox.gob.mx" }]
keywords = ["open edx", "oauth", "oauth2", "llavemx", "mexico", "identity"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "social-auth-core>=4.3.0",
  "social-auth-app-django>=5.0.0"
]

[project.urls]
Homepage = "https://github.com/aprendemx/Edx-Oauth2"
Issues = "https://github.com/aprendemx/Edx-Oauth2/issues"

[tool.hatch.build.targets.wheel]
packages = ["oauth2_llavemx"]
